Decouple the board for different games
